put the database call outside the component to avoid repetitive API calls during component development
